tested all URI endpoints and fixed CRUD that was not working

+++ server/controller/user.js
@@ -1,12 +1,12 @@
 const User = require("../model/User");
 const asyncHandler = require("express-async-handler");
-const fs = require("fs")
+const fs = require("fs");
 
 /**
  * getAllUser:  RESTful GET request returning all job objects
  */
 const getAllUser = asyncHandler(async (req, res, next) => {
-  const user = await User.find({});
+  const user = await User.find({ roles: { $ne: "Admin" } });
   res.json(user);
 });
 
@@ -28,19 +28,21 @@ const updateUser = asyncHandler(async (req, res, next) => {
   const { id } = req.user;
   const { name, password, roles } = req.body;
 
-  if(req.file){
-    console.log("Not running")
-    const user = await User.findByIdAndUpdate(id, {avatar: req.file.path})
-    fs.unlinkSync(user.avatar)
-    return
-  }
-
   if (name || password || roles)
     return res.json({ msg: "Unable to update these fields" });
-  const user = await User.findByIdAndUpdate(id, req.body);
 
-  if (!user) return res.json({ msg: "No job found with that id" });
-  res.json({ msg: "Updated successfully" });
+  if (req.file) {
+    const user = await User.findByIdAndUpdate(id, {...req.body, avatar: req.file.path });
+    if (!user.avatar) {
+      fs.unlinkSync(user.avatar);
+    }
+    if (!user) return res.json({ msg: "No job found with that id" });
+    res.json({ msg: "Updated successfully" });
+  } else {
+    const user = await User.findByIdAndUpdate(id, req.body);
+    if (!user) return res.json({ msg: "No job found with that id" });
+    res.json({ msg: "Updated successfully" });
+  }
 });
 
 /**
@@ -51,9 +53,8 @@ const deleteUser = asyncHandler(async (req, res, next) => {
   // giving the admin access to id to delete user
   let { id } = req.params;
 
-  // if the admin didn't provide any id then delete the signed in id i.e. jobseeker/employer deleting their id
+  // if the admin didn't provide any id then delete the signed in user i.e. jobseeker/employer deleting their id
   if (!id) {
-    console.log("This is runing");
     id = req.user.id;
   }
